Remove input parameter DOCKERIMAGEHASH
As discussed with Felix and Ingmar:
we don't need to pass the parameter DOCKERIMAGEHASH
into the planerio-microservice-deployment-triggerDeployment lambda.

Input validation and processing logic changed:
previously, a docker image hash or a S3 object version was required
(but not both).
In fact, there is the use case where we want to deploy a lambda
into an S3 bucket, and need a docker container for the DB migrations,
so there is in fact no either/or.